+++ src/SaveBackup/docs/release-notes.md
@@ -1,6 +1,9 @@
 [← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added options for which saves to back up (survival, Wintermute, or both).
+
 ## 1.1.0
 Released 21 April 2026 for The Long Dark 2.50+.